// AngularArmetiss/src/app/gestion-produit/models/product.ts
export class Product {
  Id_Product: number;
  Product_Name: string;
  Product_Sale_Price_HTVA: number;
  Product_Sale_Price_TVAC: number;
  Product_Description: string;
  Product_Quantity: number;
  Product_Image_URL: string | null;
  Product_Visibility: number;
  Id_TVA: number;
  Id_Category : number;
}